Rework png2sn: better compression & more generic
Removed the anti-alias-removing parts, and the palette size option.
Instead, the palette is automatically generated by the colors in the
image, with a slight sensitivity to colors that are very close together
(in which case it corrects towards the most common color).

The algorithm now also optimizes the palette order, since different
palette orders result in different compression levels. There are also
slight performance improvements, although the palette ordering logic
probably overshadows that.